[#52] add pages, apply hamburger nav link
