+++ src/pages/DontWorry/Case/index.tsx
@@ -1,0 +1,7 @@
+import Nav from "src/components/Nav";
+
+function Case() {
+  return <Nav />;
+}
+
+export default Case;
