+++ src/pages/DontWorry/Law/index.tsx
@@ -1,0 +1,7 @@
+import Nav from "src/components/Nav";
+
+function Law() {
+  return <Nav />;
+}
+
+export default Law;
