+++ src/pages/ImageFolder/index.tsx
@@ -1,23 +1,22 @@
-import React from 'react';
-import ImageFOL from 'src/components/ImageFOL';
-import styled from 'styled-components';
-import Nav from 'src/components/Nav';
-import BottomNav from 'src/components/Nav/BottomNav';
+import ImageFOL from "src/components/ImageFOL";
+import Nav from "src/components/Nav";
+import BottomNav from "src/components/Nav/BottomNav";
+import styled from "styled-components";
 
 function ImageFolder() {
-    return (
-        <>
-            <Nav />
-            <StyledImageFolder>
-                <ImageFOL month="1월" quantity={13} />
-                <ImageFOL month="1월" quantity={13} />
-                <ImageFOL month="1월" quantity={13} />
-                <ImageFOL month="1월" quantity={13} />
-                <ImageFOL month="1월" quantity={13} />
-            </StyledImageFolder>
-            <BottomNav selected="Images" />
-        </>
-    );
+  return (
+    <>
+      <Nav />
+      <StyledImageFolder>
+        <ImageFOL month="1월" quantity={13} />
+        <ImageFOL month="1월" quantity={13} />
+        <ImageFOL month="1월" quantity={13} />
+        <ImageFOL month="1월" quantity={13} />
+        <ImageFOL month="1월" quantity={13} />
+      </StyledImageFolder>
+      <BottomNav selected="Images" />
+    </>
+  );
 }
 
 export default ImageFolder;
